backend/requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
sentence-transformers
pypdf
numpy
ollama